modularize orchestrator config
