+++ crates/madara/src/config/constants.rs
@@ -2,5 +2,3 @@
 pub const SEPOLIA_RPC_URL: &str = "https://ethereum-sepolia-rpc.publicnode.com";
 #[allow(unused)]
 pub const SEPOLIA_CHAIN_ID: u64 = 11155111;
-
-pub const DEFAULT_ATLANTIC_URL: &str = "https://atlantic.api.herodotus.cloud";
