+++ crates/madara/src/config/global_config.rs
@@ -1,49 +1,19 @@
-#![allow(unused)]
-use anyhow::{Context, Error};
-use clap::builder::Str;
-use cliclack::Input;
+use std::fs;
+
 use figment::{
     providers::{Format, Toml},
     Figment,
 };
-use hex;
-use madara_cli_common::{
-    logger,
-    validation::{
-        validate_eth_address, validate_filename, validate_private_key, validate_time_with_unit,
-        validate_u64, validate_url,
-    },
-    Prompt,
-};
-use std::{fs, path::PathBuf, str::FromStr};
+use madara_cli_common::{logger, validation::validate_filename, Prompt};
 
 use serde::{Deserialize, Serialize};
 
 use crate::constants::DEFAULT_LOCAL_CONFIG_FILE;
 
 use super::{
-    constants::DEFAULT_ATLANTIC_URL,
-    eth_wallet::EthWallet,
-    l1_config::{self, L1Configuration},
-    madara::MadaraConfiguration,
+    eth_wallet::EthWallet, l1_config::L1Configuration, madara::MadaraConfiguration,
+    orchestrator::OrchestratorConfiguration,
 };
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct OrchestratorConfiguration {
-    pub atlantic_service_url: String,
-    pub minimum_block_to_process: u64,
-    pub maximum_block_to_process: Option<u64>,
-}
-
-impl Default for OrchestratorConfiguration {
-    fn default() -> Self {
-        Self {
-            atlantic_service_url: DEFAULT_ATLANTIC_URL.to_string(),
-            minimum_block_to_process: 1,
-            maximum_block_to_process: Some(100),
-        }
-    }
-}
 
 #[derive(Default, Debug, Serialize, Deserialize, Clone)]
 pub struct Config {
@@ -87,27 +57,7 @@ impl Config {
         MadaraConfiguration::init(&mut local_template)?;
 
         // Orchestrator configuration
-        logger::new_empty_line();
-        logger::note(
-            "Orchestrator configuration",
-            "You'll need to setup all the parameters related to Orchestrator",
-        );
-        let atlantic_url =
-            Prompt::new("Enter the Atlantic prover URL (e.g., http://localhost:8080)")
-                .default(&local_template.orchestrator.atlantic_service_url)
-                .validate_interactively(validate_url)
-                .ask();
-
-        let maximum_block_to_process: Option<u64> =
-            Prompt::new("Enter the maximum block to process (leave empty for no limit)")
-                .allow_empty()
-                .validate_interactively(validate_u64)
-                .ask::<String>()
-                .parse()
-                .ok();
-
-        local_template.orchestrator.atlantic_service_url = atlantic_url;
-        local_template.orchestrator.maximum_block_to_process = maximum_block_to_process;
+        OrchestratorConfiguration::init(&mut local_template)?;
 
         local_template.save(&format!("deps/data/{}", config_file_name));
         Ok(())
