+++ crates/madara/src/config/mod.rs
@@ -5,5 +5,6 @@ pub mod madara;
 
 mod eth_wallet;
 mod l1_config;
+mod orchestrator;
 
 pub use eth_wallet::EthWallet;
